requirements.txt
scikit-learn==0.23.2
streamlit==0.70.0
seaborn==0.9.0
numpy==1.19.3
matplotlib==3.0.3
pandas==0.24.2
Pillow==8.0.1